% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst-class.R
\name{as.data.frame.sunburst_data}
\alias{as.data.frame.sunburst_data}
\title{Convert sunburst_data to data.frame}
\usage{
\method{as.data.frame}{sunburst_data}(x, ...)
}
\arguments{
\item{x}{A \code{sunburst_data} object.}

\item{...}{Ignored.}
}
\value{
A data.frame containing rectangle coordinates.
}
\description{
Returns the rectangle data (\verb{$rects}).
}
