% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst-class.R
\name{$.sunburst_data}
\alias{$.sunburst_data}
\title{Access sunburst_data components}
\usage{
\method{$}{sunburst_data}(x, name)
}
\arguments{
\item{x}{A \code{sunburst_data} object.}

\item{name}{Component name.}
}
\description{
The \verb{$data} accessor is an alias for \verb{$rects}.
}
