% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst-class.R
\name{print.sunburst_data}
\alias{print.sunburst_data}
\title{Print a sunburst_data object}
\usage{
\method{print}{sunburst_data}(x, ...)
}
\arguments{
\item{x}{A \code{sunburst_data} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Displays a concise summary using \code{cli} formatting.
}
